refactoring for Key-, Group-, and CollectionConfig classes
