+++ models/DataObject/Classificationstore/CollectionConfig.php
@@ -25,10 +25,7 @@ use Pimcore\Model;
  */
 final class CollectionConfig extends Model\AbstractModel
 {
-    /**
-     * @var bool
-     */
-    public static $cacheEnabled = false;
+    use Cache\RuntimeCacheTrait;
 
     /**
      * @var int|null
@@ -129,22 +126,6 @@ final class CollectionConfig extends Model\AbstractModel
         $config->save();
 
         return $config;
-    }
-
-    /**
-     * @param bool $cacheEnabled
-     */
-    public static function setCacheEnabled($cacheEnabled)
-    {
-        self::$cacheEnabled = $cacheEnabled;
-    }
-
-    /**
-     * @return bool
-     */
-    public static function getCacheEnabled()
-    {
-        return self::$cacheEnabled;
     }
 
     /**
@@ -295,20 +276,6 @@ final class CollectionConfig extends Model\AbstractModel
     }
 
     /**
-     * Returns all group belonging to this collection
-     *
-     * @return CollectionGroupRelation[]
-     */
-    public function getRelations()
-    {
-        $list = new CollectionGroupRelation\Listing();
-        $list->setCondition('colId = ' . $this->id);
-        $list = $list->load();
-
-        return $list;
-    }
-
-    /**
      * @return int
      */
     public function getStoreId()
@@ -322,48 +289,6 @@ final class CollectionConfig extends Model\AbstractModel
     public function setStoreId($storeId)
     {
         $this->storeId = $storeId;
-    }
-
-    /**
-     * Set cache item for a given cache key
-     *
-     * @param CollectionConfig $config
-     * @param string $cacheKey
-     */
-    private static function setCache(CollectionConfig $config, string $cacheKey): void
-    {
-        if (self::$cacheEnabled) {
-            Cache\Runtime::set($cacheKey, $config);
-        }
-
-        Cache::save($config, $cacheKey, [], null, 0, true);
-    }
-
-    /**
-     * Remove a cache item for a given cache key
-     *
-     * @param string $cacheKey
-     */
-    private static function removeCache(string $cacheKey): void
-    {
-        Cache::remove($cacheKey);
-        Cache\Runtime::set($cacheKey, null);
-    }
-
-    /**
-     * Get a cache item for a given cache key
-     *
-     * @param string $cacheKey
-     *
-     * @return CollectionConfig|bool
-     */
-    private static function getCache(string $cacheKey): CollectionConfig|bool
-    {
-        if (self::$cacheEnabled && Cache\Runtime::isRegistered($cacheKey) && $config = Cache\Runtime::get($cacheKey)) {
-            return $config;
-        }
-
-        return Cache::load($cacheKey);
     }
 
     /**
