+++ models/DataObject/Classificationstore/KeyConfig.php
@@ -25,10 +25,7 @@ use Pimcore\Model;
  */
 final class KeyConfig extends Model\AbstractModel
 {
-    /**
-     * @var bool
-     */
-    public static $cacheEnabled = false;
+    use Cache\RuntimeCacheTrait;
 
     /**
      * @var int|null
@@ -149,22 +146,6 @@ final class KeyConfig extends Model\AbstractModel
         $config->save();
 
         return $config;
-    }
-
-    /**
-     * @param bool $cacheEnabled
-     */
-    public static function setCacheEnabled($cacheEnabled)
-    {
-        self::$cacheEnabled = $cacheEnabled;
-    }
-
-    /**
-     * @return bool
-     */
-    public static function getCacheEnabled()
-    {
-        return self::$cacheEnabled;
     }
 
     /**
@@ -396,52 +377,10 @@ final class KeyConfig extends Model\AbstractModel
     }
 
     /**
-     * Set cache item for a given cache key
-     *
-     * @param KeyConfig $config
-     * @param string $cacheKey
-     */
-    private static function setCache(KeyConfig $config, string $cacheKey): void
-    {
-        if (self::$cacheEnabled) {
-            Cache\Runtime::set($cacheKey, $config);
-        }
-
-        Cache::save($config, $cacheKey, [], null, 0, true);
-    }
-
-    /**
-     * Remove a cache item for a given cache key
-     *
-     * @param string $cacheKey
-     */
-    private static function removeCache(string $cacheKey): void
-    {
-        Cache::remove($cacheKey);
-        Cache\Runtime::set($cacheKey, null);
-    }
-
-    /**
-     * Get a cache item for a given cache key
-     *
-     * @param string $cacheKey
-     *
-     * @return KeyConfig|bool
-     */
-    private static function getCache(string $cacheKey): KeyConfig|bool
-    {
-        if (self::$cacheEnabled && Cache\Runtime::isRegistered($cacheKey) && $config = Cache\Runtime::get($cacheKey)) {
-            return $config;
-        }
-
-        return Cache::load($cacheKey);
-    }
-
-    /**
      * Calculate cache key
      *
      * @param int $id
-     * @param string|null $name
+     * @param string $name
      *
      * @return string
      */
